Refactor project layout and add makefile
