+++ .gitignore
@@ -1,2 +1,5 @@
 .vscode/
 .idea/
+bin/
+dep/
+obj/
