+++ doc/PROPOSAL.md
