--- src/fragment.c
@@ -1,4 +1,0 @@
-//! @file fragment.c
-//!
-//! Splits large events (which exceed the maximum size limit of the database) 
-//! into multiple fragments for storage in FoundationDB key/value pairs.
